Safari-609.1.11.1/LayoutTests/fast/dom/nodeListIterator-expected.txt

Testing of NodeList iterable capacities.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testedIterable.entries is entriesFunction
PASS testedIterable.forEach is forEachFunction
PASS testedIterable.keys is keysFunction
PASS testedIterable.values is valuesFunction
PASS testedIterable.length is 2
PASS item is children[index++]
PASS item is children[index++]
PASS pair[0] is children[0]
PASS pair[1] is children[1]
PASS forEachContainer is testedIterable
PASS forEachIndex is index
PASS node is children[index++]
PASS thisValue is window
PASS forEachContainer is testedIterable
PASS forEachIndex is index
PASS node is children[index++]
PASS thisValue is window
PASS thisValue is window
PASS thisValue is window
PASS thisValue is givenThisValue
PASS thisValue is givenThisValue
PASS iterator.next().value is 0
PASS iterator.next().value is 1
PASS end.done is true
PASS end.value is undefined.
PASS iterator.next().value is children[0]
PASS iterator.next().value is children[1]
PASS end.done is true
PASS end.value is undefined.
PASS pair.length is 2
PASS pair[0] is 0
PASS pair[1] is children[0]
PASS pair.length is 2
PASS pair[0] is 1
PASS pair[1] is children[1]
PASS end.done is true
PASS end.value is undefined.
PASS end.done is true
PASS end.value is undefined.
PASS descriptor.configurable is true
PASS descriptor.writable is true
PASS descriptor.enumerable is false
PASS testedIterablePrototype[Symbol.iterator] = valuesFunction; did not throw exception.
PASS checkItemType(a) is true
PASS checkItemType(a) is true
PASS checkItemType(a) is true
PASS checkItemType(a) is true
PASS counter is 4
PASS checkItemType(v) is true
PASS checkItemType(v) is true
PASS checkItemType(v) is true
PASS checkItemType(v) is true
PASS counter is 4
PASS typeof k is "number"
PASS typeof k is "number"
PASS typeof k is "number"
PASS typeof k is "number"
PASS counter is 4
PASS typeof e[0] is "number"
PASS e[1] is testedIterable[e[0]]
PASS typeof e[0] is "number"
PASS e[1] is testedIterable[e[0]]
PASS typeof e[0] is "number"
PASS e[1] is testedIterable[e[0]]
PASS typeof e[0] is "number"
PASS e[1] is testedIterable[e[0]]
PASS counter is 4
PASS successfullyParsed is true

TEST COMPLETE
